+++ tutorials/Tutorial5_Evaluation.py
@@ -5,13 +5,12 @@ from haystack.retriever.dense import DensePassageRetriever
 from haystack.eval import EvalReader, EvalRetriever
 from haystack.reader.farm import FARMReader
 from haystack.preprocessor import PreProcessor
+from haystack.utils import launch_es
 from haystack import Pipeline
 
 from farm.utils import initialize_device_settings
 
 import logging
-import subprocess
-import time
 
 logger = logging.getLogger(__name__)
 
@@ -153,6 +152,7 @@ def tutorial5_evaluation():
             )
             results.append(res)
 
+        n_queries = len(labels)
         eval_retriever.print()
         print()
         retriever.print_time()
@@ -161,23 +161,8 @@ def tutorial5_evaluation():
         print()
         reader.print_time()
         print()
-        eval_reader.print(mode="pipeline")
+        eval_reader.print(mode="pipeline", n_queries=n_queries)
 
-
-def launch_es():
-    # Start an Elasticsearch server
-    # You can start Elasticsearch on your local machine instance using Docker. If Docker is not readily available in
-    # your environment (eg., in Colab notebooks), then you can manually download and execute Elasticsearch from source.
-
-    logger.info("Starting Elasticsearch ...")
-    status = subprocess.run(
-        ['docker run -d -p 9200:9200 -e "discovery.type=single-node" elasticsearch:7.9.2'], shell=True
-    )
-    if status.returncode:
-        logger.warning("Tried to start Elasticsearch through Docker but this failed. "
-                       "It is likely that there is already an existing Elasticsearch instance running. ")
-    else:
-        time.sleep(15)
 
 if __name__ == "__main__":
     tutorial5_evaluation()
